% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{f}
\alias{f}
\title{simple functio}
\usage{
f(x)
}
\arguments{
\item{x}{x is a number}
}
\value{
a dataframe
}
\description{
today, I create my first function, a very useful function.
}
\details{
This is some description of this function.

you can use this function to calculate x+1, then return the value of x+1
}
